fix: move the test to backend-admin folder
