+++ backend-admin/Jenkinsfile
@@ -1,0 +1,39 @@
+// Expected parameters for the pipeline:
+//   - OPENSHIFT_URL - RHMI cluster to target (https://...)
+//   - OPENSHIFT_ADMIN_USERNAME - Admin username
+//   - OPENSHIFT_ADMIN_PASSWORD - Admin user password
+
+pipeline {
+  agent {
+    dockerfile {
+      dir 'containers/node'
+      label 'psi_rhel8'
+    }
+  }
+  stages {
+    stage('Install dependencies') {
+      steps {
+        sh 'npm install'
+        dir('backend') {
+          sh 'npm install'
+        }
+      }
+    }
+    stage('Login with oc') {
+      steps {
+        wrap([$class: 'MaskPasswordsBuildWrapper', varPasswordPairs: [[var: 'OPENSHIFT_ADMIN_PASSWORD', password: 'OPENSHIFT_ADMIN_PASSWORD']]]) {
+          sh "oc login $OPENSHIFT_URL -u $OPENSHIFT_ADMIN_USERNAME -p ${OPENSHIFT_ADMIN_PASSWORD}"
+        }
+      }
+    }
+    stage('Test') {
+      steps {
+        dir('backend') {
+          sh 'npm start -- --reporter mocha-jenkins-reporter || true'
+          archiveArtifacts JUNIT_REPORT_PATH
+          junit allowEmptyResults: true, testResults: JUNIT_REPORT_PATH
+        }
+      }
+    }
+  }
+}
